+++ projects/batik-1.19/src/Entrypoint.java
@@ -9,70 +9,65 @@ import org.apache.batik.apps.rasterizer.SVGConverter;
 import org.apache.batik.apps.rasterizer.DestinationType;
 
 public class Entrypoint {
+
+    private static File dst;
     public static void entrypoint(Path svg) {
         try {
-            File dst = Files.createTempDirectory("output").toFile();
+            SVGConverter converter = new SVGConverter();
+            converter.setSources(new String[]{svg.toAbsolutePath().toString()});
+            converter.setDst(dst);
+
             try {
-                SVGConverter converter = new SVGConverter();
-                converter.setSources(new String[]{svg.toAbsolutePath().toString()});
-                converter.setDst(dst);
-
-                try {
-                    converter.execute();
-                } catch(Exception exc) {
-                    exc.printStackTrace(System.err);
-                }
-
-                try {
-                    converter.setDestinationType(DestinationType.JPEG);
-                    converter.execute();
-                } catch(Exception exc) {
-                    exc.printStackTrace(System.err);
-                }
-
-                try {
-                    converter.setDestinationType(DestinationType.TIFF);
-                    converter.execute();
-                } catch(Exception exc) {
-                    exc.printStackTrace(System.err);
-                }
-
-                try {
-                    converter.setDestinationType(DestinationType.PDF);
-                    converter.execute();
-                } catch(Exception exc) {
-                    exc.printStackTrace(System.err);
-                }
-            } catch (Throwable exc) {
+                converter.execute();
+            } catch(Exception exc) {
                 exc.printStackTrace(System.err);
-            } finally {
-                deleteDirectory(dst);
+            }
+
+            try {
+                converter.setDestinationType(DestinationType.JPEG);
+                converter.execute();
+            } catch(Exception exc) {
+                exc.printStackTrace(System.err);
+            }
+
+            try {
+                converter.setDestinationType(DestinationType.TIFF);
+                converter.execute();
+            } catch(Exception exc) {
+                exc.printStackTrace(System.err);
+            }
+
+            try {
+                converter.setDestinationType(DestinationType.PDF);
+                converter.execute();
+            } catch(Exception exc) {
+                exc.printStackTrace(System.err);
             }
         } catch (Throwable exc) {
             exc.printStackTrace(System.err);
+        } finally {
+            deleteDirectoryContents(dst, true);
         }
-
     }
 
-    public static boolean deleteDirectory(File directoryToBeDeleted) {
+    public static void deleteDirectoryContents(File directoryToBeDeleted, Boolean root) {
         File[] allContents = directoryToBeDeleted.listFiles();
         if (allContents != null) {
             for (File file : allContents) {
-                deleteDirectory(file);
+                deleteDirectoryContents(file, false);
             }
         }
-        return directoryToBeDeleted.delete();
-    }
-
-    public static void recurseDirectories(Path path) throws Exception {
-        try(Stream<Path> files = Files.walk(path)) {
-            for(Path inputFile: (Iterable<Path>) files.filter(Files::isRegularFile)::iterator) {
-                Entrypoint.entrypoint(inputFile);
-            }
+        if(! root) {
+            directoryToBeDeleted.delete();
         }
     }
 
     public static void main(String args[]) throws Exception {
-        recurseDirectories(Paths.get(args[0]));
+        Entrypoint.dst = Files.createTempDirectory("dst").toFile();
+        try(Stream<Path> files = Files.walk(Paths.get(args[0]))) {
+            for(Path inputFile: (Iterable<Path>) files.filter(Files::isRegularFile)::iterator) {
+                Entrypoint.entrypoint(inputFile);
+            }
+        }
     }
 }
