+++ projects/hsqldb-2.0.0/src/Entrypoint.java
@@ -13,52 +13,40 @@ import java.sql.DriverManager;
 import org.hsqldb.cmdline.SqlFile;
 
 public class Entrypoint {
+
+    private static Path db;
     public static void entrypoint(Path input) {
-        try {
-            // Create a temporary directory for the database.
-            Path db = Files.createTempDirectory("fuzzing");
-
-            try {
-                Connection conn = DriverManager.getConnection("jdbc:hsqldb:file:" + db.toString() + "/testdb", "SA", "");
-
-                SqlFile sqlFile = new SqlFile(input.toFile());
-                Map<String, String> userVars = new HashMap<>();
-                userVars.put("*SCRIPT_DIR", input.getParent().toString());
-                sqlFile.addUserVars(userVars);
-                sqlFile.setConnection(conn);
-                sqlFile.execute();
-            } catch (Throwable exc) {
-                exc.printStackTrace(System.err);
-            } finally {
-                // Clean up the temporary files and directory
-                deleteDirectory(db.toFile());
-            }
+        try(Connection conn = DriverManager.getConnection("jdbc:hsqldb:file:" + db.toString() + "/testdb", "SA", "")) {
+            SqlFile sqlFile = new SqlFile(input.toFile());
+            Map<String, String> userVars = new HashMap<>();
+            userVars.put("*SCRIPT_DIR", input.getParent().toString());
+            sqlFile.addUserVars(userVars);
+            sqlFile.setConnection(conn);
+            sqlFile.execute();
         } catch (Throwable exc) {
             exc.printStackTrace(System.err);
+        } finally {
+            deleteDirectoryContents(db.toFile(), true);
         }
     }
 
-
-    public static void recurseDirectories(Path path) throws Exception {
-        try(Stream<Path> files = Files.walk(path)) {
-            for(Path inputPath: (Iterable<Path>) files.filter(Files::isRegularFile)::iterator) {
-                Entrypoint.entrypoint(inputPath);
-            }
-        }
-    }
-
-    public static boolean deleteDirectory(File directoryToBeDeleted) {
+    public static void deleteDirectoryContents(File directoryToBeDeleted, Boolean root) {
         File[] allContents = directoryToBeDeleted.listFiles();
         if (allContents != null) {
             for (File file : allContents) {
-                deleteDirectory(file);
+                deleteDirectoryContents(file, false);
             }
         }
-        return directoryToBeDeleted.delete();
+        if(! root) {
+            directoryToBeDeleted.delete();
+        }
     }
-
     public static void main(String[] args) throws Exception {
-        // Calls `recurseDirectories` with the path provided as a command-line argument
-        recurseDirectories(Paths.get(args[0]));
+        Entrypoint.db = Files.createTempDirectory("db");
+        try(Stream<Path> files = Files.walk(Paths.get(args[0]))) {
+            for(Path inputFile: (Iterable<Path>) files.filter(Files::isRegularFile)::iterator) {
+                Entrypoint.entrypoint(inputFile);
+            }
+        }
     }
 }
